Added AI and balls for it to follow
Added the AI I made a couple of days ago, and also the changes I've been
trying out. There are now a ring of spheres around the player, which the
AI will, or should, try to target in order to simulate driving around
the player.

This workes ish, but I need an actual ship to test it with. Will work
more with it tomorrow.