AI Targeting and AI spawning
There's a sircle of balls around the player. The AI now follows these
instead of the player. This is now working, but the function is open for
tweeking. The balls are put into an array in the "MoveTo" script on the
ship with the "GameObject.FindAllObjectsWithTag("playerBalls");.
Literally one line of code.

The AI also spawns on specific points now. These points are used to
determine where the AI should spawn, and what "lvl" it is. The 3 squares
to the left spawns the lvl 1 ship, and the 3 to the right spawns the lvl
2 ship. The difference in difficulty is only that the lvl2 ship deals 2
damage instead of 1 per cannonball.

Because of this I had to make a prefab for both ships and one prefab for
the lvl 1 cannonball and one for the lvl 2 cannonball. There are no
difference in the scripts, but the tags are different. This is handled
by a "GM" object in the scene., which holds the "spawnAI" script.