Cleaned my project and scripts
I've made prefabs of all of my objects, and I've commented my code. I
also deleted som old scripts that are no longer in use, and some that
never was used at all. That is mostly what I've done today. I also
tested the AI with Lars's iteration of the playercontrol, and it seemes
to work just fine.