AI vs Gravitation
Added gravity to the scene to test out how the AI will interact with it.
Result? Not so good, but at the same time as expected.

The AI rammed right into it, ignoring the danger.

I've now added new path-rules for it to follow. They have path cost 2, 4
and 6, in order to who is closest and furthest from the planet. This
works if the player is on the other side of the planet, but if the AI is
right behind the player it will ignore the new paths and drive straight
for it.