Objects follows Agent!
I've now done it so that the Agent, aka AI, now is a cylinder instead of
the ship itself. The ship will now follow the Agent, and the Agent will
follow the player balls. This means that the AI Ship now can have custom
made boundaries for the movement, which we now got. The AI ship has the
same constraint as the player ship, giving it a much smoother and more
fair balancing.

When the player ship stops the AI will also act differently. The Agent
will go to one of the balls right next to the player, depending on which
one is closest to the AI ship. The AI Ship will den drive towards the
Agent, and when they hittest the ship will stop moving forwards, and
also turn to face its canons directly at the Player ship.

AI no longer spawns. I have 1 AI in the scene at any given moment, so
I'll have to fix this at a later time when the AI is in the state we
want it to be.