Player balls, AI can die and respawn, fixed colliders and re-wrote a shitton of code
Made the balls around the player separate objects in the hierarchy, and
also so that the follow the player, This means that the balls won't
rotate when the player does, making it easier to make sure the AI Ship
stands still while the player ship rotates without driving forwards.
This is a step forward in making an AI that seems intelligent and.

The AI can now die and respawn now, making them an aqual functional
feature in the game. AI_LVL1 and AI_LVL2 has separate amount of health,
increasing the difficulty between them. To make the AI able to die, I
had to fix the meshcollider on the object. This was done by removing
every collider on every child of the object. I did the same thing for
the player.

I also re-wrote a lot of code. I made some functions, then I made some
better ones and deleted the one I made just before. I also cleaned up in
the scripts generally, as Lars gave me the new PlayerMove script where
the bools "turnLeft", "turnRight" and "moveForward" were public.
This allowed me to remove a lot of the calculations I had to make,
increasing efficiency and decreasing the amount of lines with code.