AI Movement and Gravity
Today I tweaked how the AI Ship acts around the player when it moves,
and when its stationary (primarily stationary). I made sure that the
ship rotates the shortest way to aim it's cannons towards the player,
and I also tried to make it so that when the player ship moves, and
stoppes again, the AI will find a ball to move to that is both closest
to the ai ship and minimum X distance between itself and the ai ship. I
did this to prevent the AI to continuesly rotate around the same point
trying to reach a point which it does not need to reach.

I also tried to start on how the AI should act around the gravity, but I
had next to no motivation and inspiration today. Will work more AI vs
Gravity later.