AI - avoiding planet and chasing player
I've made a new script where I try to handle how the AI should act when
it faces a planet. As of now there are 400+ lines of code trying to
tackle this, but because of the "inverted" AI model the tests will yield
fake results. Will try to fix the model tomorrow.

The AI checks if the player is to the front or right of itself, and if
its on the front or rear side of itself. Does the same for the planet.
When it knows where the player and planet is, it should run through a
pretty big check to control what it should do next. There are bound to
be issues in this code.