Bugfixing, new AI ship model, new AI respawn, updated AI behaviour around planet,
Fixed some bugs in the code for how the AI should behave around the
planet, and also implementet how the AI should act when the player is to
the left of it while also being close to the planet. As of now the code
is functional. It's not good, but it's not terrible either. The AI will
drive towards the player while trying to avoid the planet.

I changed the AI model to be the one of the player. This is because the
pivot point on the old AI model was inverted, messing up every
side-tests I did in my code. The new model fixes these issues.

How the AI spawns is rewamped to be more dynamic. spawnAI will create a
spawnpoint somewhere around the player, random to a certain degree, and
determins wheter or not to spawn the lvl1 or lvl2 AI based on the
distance between the player and an "origin" object. The origin object is
a cube placed in the world, simulating the starting position of the
player.