New pathfinding, new movement rules,, updated projectile stats and fleeing
I've given up on tring to write the entire behaviour of the AI from the
bottom. Instead I took inspiration from Sid Meiers Pirate's, where the
AI always drives forwards and turns based on what broadside is closest
to the player. This means that the "chasePlayer.cs" no longer is used,
and I went back to "aiMove.cs".

Our AI now does the same. It always drives forwards, and turns based on
what side of it the AI is. This works smoothly. I also changed the
movement of the AI to be equal to the player, meaning that now the AI
also uses addForce to move around.
I tested out how the AI would act while around a planet, using the old
avoidPlante script. It's not perfect, but it works, sort of. Some of the
decisions the AI makes are unlogical and should be improved, but for now
it's good enough.

The AI now have a 10% chance to initiate fleeing when it's health is
below 20%. I've yet to actualy try this out, but the code should be
solid.

I've also updated the projectile of the player to use addForce.