AI - Following player, actnig around planets, respawn and "newborn" ships
The code for following the player is complete (I think). After talking
with Ole Flaten it was confirmed that the new aproch is indeed a good
one.

How the ai act around planets have been heavily tweeked, but not enough.
I try to make it so that the ai will focus on eskaping the planet when
the planet is too close, and aim for the player when the planet is not
an imidiate danger.
I'm trying to use Raycast to detect the planet. This works, sort of, but
I might need more raycasts. At the moment I have one in the front and
one on each side. The problem now is that the planet might actually come
between the casts, causing the ai not to see the planet.

I've also tweeked the respawn function. It's nothing much, really. I
just changed how it decides where the AI should spawn.

The big change in the respawning is how newbord ships are handled. They
now have increased peed and handling until they come close enough to the
player. This ensures us that the AI will always reach the player when
they spawn, removing the fear of "AI tries to pursue the player,
resulting in an everlasting chase".