Minor AI tweeks and scraps
Made some minor changes to how the AI acts when it's to close to a
planet. It now checks the player position and takes a turn based on
that, whilst it before took decided this based on the planet position.

Scraps spawn now! They have some force when spawned, but it will quickly
decrease due to drag and angular drag. It also got some random rotation
just too look fancy. When an AI is killed there will spawn 1 up to 6
scrap items on the AI position. When the player drives close to these
there is a magnet effect dragging the scraps towards the player.
Once picked up, the player recieves a random cash reward. The random
reward is calculated in  a script the scrap object itself holds.