AI Boss and bombs!
The Boss will now spawn if the player is close enough, and boy is he a
vicious one! He has 50 health, got the same model as the player, and
spawns bombs behind him.

Every 5 seconds the boss will deploy one of his bombs. The spawn right
behind him, but far enough away so that they do not trigger instantly.
The bombs trigger if they hit a ship or canonballs.

When an object hits a bomb an explotion is created. Every bomb holds the
"bombHitSomething" script, and it is this one that checks for colitions.

When a colition happens the script will search for every collider within
a range of X, and check if the object with the collider has a Rigidbody
component. If it does not have this component, the script will check if
the root object has it (like every ship does).
A force is applied, and the objects will be moved away from the
explotion. I also added so that both the player and the AI loses
controll over their ship for 1 second after being affected by an
explotion.

The boss holds the same scripts as the regular Marine AI, meaning it
holds the same constraints for when it can shoot, when it should chase
and when it should flee. Obviously we dont want the boss to flee in the
final build, but it's ok for now.