Smoother AI movement, deleted some unused scripts
The AI now turnes to left or right based on the Raycast checks on
AIsideCanons.cs instead of checking the relative z position between
itself and the player. This results in a much smoother turning, as we
now turn whenever the canons cant shoot at the player, instead of just
always when the relative Z position between the player and AI is != 0

Also removed some unused scripts. These are ballsPosition, detectShip,
MoveTo and playerMovement. They were all used in previous versions of
the AI, but they have been replaced with newer and better scripts.