Random weapon, damage output, fixes after merge
The AI now spawns with a random ammount of upgraded cannons, and the
cannons gain either lvl 1, 2 or 3. The cannonballs they fire gain
dmageOutput based on the cannon they spawn on. lvl 1 cannon = 1 damage,
lvl 2 cannon = 3 damage and lvl 3 cannon = 5 damage.

spawnAI calculated the distance between player and originalSpawn, and
takes this number and multiplies it with 0.1, giving us the 10% number
of the distance. Then it takes Random.Range(0, 101) to recieve a random
number between 0 and 100. If the number is lover than 10% of distance,
the cannon is upgraded to lvl2.
Then it again takes 50% of that number again, runs another random test,
and if the new random number is lower than 50%, then the cannon is lvl3.