Cleaned hierarchy, random weapon upgrade, new material for cannonballs, rotating cogs
I cleaned the hierarchy, meaning I renamed some of the objects. It was
pretty hard for others to know the difference between "sidecannons" and
"sidecannons(1)", so something had to be done.
I also renmed "AI_LVL1" to "Marine".

There are now 3 new prefabs, representing 3 new cannonballs for the 3
new cannons. If a cannon is lvl 1, it will instantiate the lvl1
cannonball. So lvl2 cannon instantiates the lvl2 ball, and cannon lvl3
the lvl3 ball.
They have different colors to signefy the damage they deal. Yellow means
1 damage, blue is 3 and red is 5.

The cogs on the Marine also rotates now. Dunno how much of that you can
see in game, but they rotate none the less.