Lars's scene in my scene now, modified AI spawn, marine material changes
I implementet Lars's scene into my, so now I can actually see how the AI
behave around moving planets with gravity (yet to be done). In the
procedure I modified the spawn script, as the new lvl is far bigger than
the one I've worked with.

The Marine also changes material when damage is dealth upon it. When it
has 66% health left, it changes to the mildly damaged material, and when
its at 33% it changes the material to badly damaged.