Multiple marines, avoids shops and "sometimes" other marines, fixed spawning, patroling
There are now multiple Marine AI on the screen at the same time! This
allows for a much more vivid enviroment and a more lifelike universe. Is
it without flaws? Absolutely not!

The AI fails to avoid eachothers, and I've no idea why. According to
debugging they see eachothers, but they don't act like they should. They
can avoid shops, and use basically the same code for avoiding other
marines.

The AI also act independently, meaning that a new instantiation of a
prefab does not cause the ones already in the scene to grind to a halt.
This is obviously a necessity, and something we had to do.

Something that was also added is patroling. The AI now gets a randomly
generated position in the world to drive towards. This works by creating
a sphere at the vector point created in SpawnAI. The sphere is also
invisible. The AIPatroling script then drives towards the object exactly
like it does towards the player in AIMove.
The AI also spots the player and will stop the patroling and go in for
the attack when close enough (as of now, within 100 meters).