New unused script, killing all AI's via AIMaster and moved over to list (bad choice)
I made the bad decision to jump over to using lists instead of arrays to
hold my AI's. I'll go back to using an array tomorrow. The plan is to
kill all Marines that is not engaging in a fight against the AI. We want
there to be many ships on the scene, but only 1 in combat at any given
time.