Marines behaviour, Boss behaviour, fixed bombHit is fixed
The Marine will now flee when the boss is spawned. I did not want them
to just despawn, as we might want to spawn the boss at a distance from
the player. The Marines also don't spawn when the boss is alive or when
a fight is going on.

The boss is now almost done, I only need the model and materials. The
bombs can actually impact it, I just had to change from "GetComponent"
to "GetComponentInParent" to make sure that the script actually accessed
the root object to find AIMaster. I also added the player materials to
make sure that it could change materials when it's health is low enough.
The boss will spawn with 100 health, and still when the player is within
100m to the spawn object. These are variables that can be changed at any
time dring the balancing.
I had to make a new bool  in AIMove to make sure that the boss doesn't
despawn every single Marine out there,

There is a new function out there to make sure that a fight occures
every X second/minute. As of now it's set at 5 seconds for debug
reasons, but I didn't have time to test it out today.

Also, for some reason, every single fucking Marine goes in to killer
mode and drives towards the Player the second they spawn. I have no idea
why.