Fixed respawn in Tutorial, removed unused lines of code, remade Player prefab to be able to build again
Added the deathscreen to the tutorial so that the code would not flip a
table, removed some unused lines to clean the code and remade the prefab
for the player. Apparently something had gone wrong, causing the
player's body not to render after the game was built via Unity. Breaking
and remaking the prefab fixed this error.