Fixed an error in GameControl(load), AI can shoot again
There was an error in GameControl causing the game to recieve an error
in the Load function. The issue was that it tried to access Rigidbody on
a child, when its on the root object.

The marine and boss now spawns higher in the world (y-axis) so that the
raycast can see the player. Thus, the ships can shoot again.