Fixed GameControl bug, fixed spawnAI.spawn issues in AIMaster, player can now exit spawn
In GameControl there was an error causing the script to change the
position of a child object in Player, instead of the whole object. This
is now fixed, so that it's the entire object that is modified.

I also removed "private spawnAI spawn" from AIMaster and moved the
declaration over to spawnAI.cs. This seems to have fixed the issue with
"not set to an instance of an gameobject" (or something like that) error
that would sometimes occur.

I also increased the number Vector3.forward is multiplied with. It was
to small before, so the player would spawn inside the shop when entering
the World scene again.