NO RANDOM CRASHES, changed what marine that attacks after X seconds
I found the issue! When a marine is spawned, they gain weapons with a
random level. In the script as of now, there is a check to see if at
least 2 cannons are upgraded when the marine is more than 200meters away
from Origin. If there are only 1 og 0 cannons upgraded, it will run a
"while" test untill enough cannons are upgraded. It does so by checking
a bool array to see if any index is false, meaning there is an available
stop for a canon to be upgraded.
Here is the issue: I forgot to clean the array after spawning a marine.
This made it so that after 6 spawns every bool in the array said "true",
meaning that the cannon was already upgraded. I fixed this by creating a
for loop in AIsideCanons that resets the array after cloning the data
from it.

I also changed it to be the marine that is closest to the player that
goes into attack mode  after X seconds. It used to be the last spawned
ship, but this was stupid. Sometimes one ship would be close to the
player, whilst the last spawned ship would be driving in the complete
oposite direction of the player.