General AI changes, increased mass of scraps, modified cannon upgrades
I made some minor changes to the AI based on the feedback we got during
yesterdays alpha testing. Some of the changes are also made after the
brutal misuse John gave my lovely AI.
-Marine now despawns when distance to player is greater than 300, than
previously 500.
-Marine has increased movement speed (force) until it has catched up to
the player. 1000 force until distance is 60, then it is 650.
-Marine doesn't detect player before distance between them is less than
150. Previously it was 1000.

A lot of people disliked the fact that scraps would get devoured by the
planets nearly all of the time, due to the fact that a planet is close
to the spawn. I increased the mass of "scraps" by 1000, whilst it
previously was 0.1. I also increased the force they gain at spawn to
compensate for the increased mass, so that they still move a little away
from eachother.

Previously the chances of an AI's cannons level was 10% of the distance
between player and origin. This ment that when the player was 1000m from
spawn, there would already be 100% chance of every cannon to be level 2.
This was infact way to fast.
The new number is 2% of distance. Now the distance must be 5000m for it
to be 100% chance for all to be level 2 cannons.