Cargo ship is implemented
Every 10 seconds the code will check and see if there is a living Cargo
ship. If there isn't, it will spawn one directly in front of the player.
The Cargo ship will flee from the player when its too close, and it will
cause the marines to despawn at the same time this happens.
When the Cargo ship is down to 66% remaining health, it will go into
attack state. When the Cargo ship dies, it spawns quite a lot more
scraps than the Marine does.