Modified how the AI detects hinderings in the world
Instead of 1 RayCast to each side, every AI now has 2 casts to the right
and 2 to the left. One at the tip of the ship, and one at the rear. This
should improve how easy it is for the AI to detect planets, shops and
other ai's to avoid. This is done by creating 2 boxes that are children
of the object, and they cast their own left and right RayCast. If either
cast hit something, we check what it is and return that value to
AIAvoid.cs.