Explotion effect on cannonball hit
There will now play an explotion when a cannonball hit something and
deletes itself. This is part of the "ToonPack" partciles that Sunniva or
Joakim bought, so I just implemented one of the prefabs from that asset.
When a cannonball hit something it will Instantiate a copy of the preab,
then place the prefab at its on location. On the explotion prefab there
is a script that deletes itself and the object after 2 seconds.