Smoke Particles when AI at 66% health
When the AI is at 66% health left, smoke particles will start playing.
These are placed on the marine, and will play at those positions. I put
them in an array in DamagedAI.cs, where I first deactivate them, then I
got a function to activate them.

I made this for the player also, but I can't access GameControl, as this
is John's script.