Minimap and Cargo ship bugfix
There was an error when the cargo ship spawned. This was because it
would spawn and instantly try to flee, even before it had saved all the
data it needed. It no longer spawns and instantly try to flee.

I downloaded a free asset from the Unity store, where the asset would
calculate what objects should appear on the radar, the size of the
radar, size of the objects on the radar and so on. It's pretty amazing.