Minimap implemented!
A functional minimap with zoom is now in the game. It will detect all
types of enemies and shops, and the player will be able to zoom in and
out with the buttons under the minimap. When the player press a button
the icons will also scale, but for now the scale ammount it pretty much
just a random number that seems fitting. I tried a couple of scale
numbers, and zoom out got 0.9 and zoom in got 1.111. These number looked
pretty good.

I had to make sure that newly spawned enemies got the correct scale, so
I had to make some changes to the "Mapmarker" script that came with the
asset I used. It checks if the object its sitting on is with the tag
"aiShip", and if it is it will get the same scale that zoomMinimap holds
in the variable "enemyScale".
New ships also needs to check in the update function if their scale is
!= to enemyScale. If this is true, then we just change the scale to
enemyScale.