New boss
The new boss is finally here! With it's shiny model and fucking 12 lvl 3
cannons he is going to fuck shit up. Literally. He is beyond OP.

The boss works exactly like the previous boss, only that now he seems
intimidating. He is big. He is blue. He is scary. He also has 2 extra
ray casts to detect if the cannons can shoot at the player, and this is
1 of 2 major changes.
"RayCastPlayer" is a script set on 3 objects that are put on the Boss
object. These sends our ray casts to the left and right, detecting the
player. When they detect the player they say that to AIsideCanons, and
that script will shoot.

The other major change is the scale of the Boss. It is huge compared to
anything else in the game. And I love that, I truly do.