Bugfixes (ball explotions, spawning behaviour, checks if its the tutorial or not)
The spawn position of the explotion that occures when a cannonball hit
something has been improved. Previously we would spawn the explotion,
and then change the position. This caused a bug where an explotion would
only play at the next cannonball hit. I.e. cannonball A hits something,
but the explotion will only spawn when cannonball B hits something.

When the boss was alive other AI's would still spawn around the player.
This is not ideal, as we want the fight to be hard and epic. This is now
fixed by checking if there is a "Boss(Clone)" active in the hierarchy
before the spawnShip() function is run.
There was also an issue where the marines would not start respawning
after a cargo ship in combat died. This was simply because I forgot to
set "stopSpawn" and "stopFightTimer" to false.

Both AIMaster and ScrapRandomDirection are all checking to see if the
game is in the Turorial scene or not. This is because they are used
during the tutorial, and certain checks needs to be done. We had to
change the if-test to search for something else than "is the Tutorial
script on the TutorialContrl object active and enabled?" because we
deleted the object from the World scene hierarchy. So now we test if the
TutorialControl object exist or not. If does not exist, we ignore the
code.