Improved spawing of Marine and Cargo
They will no longer spawn completely randomly, as they now have certain
rules to follow. These are only that before they spawn, the code will
check if there are any other objects within a 10 meter radius around the
newly created spawnpoint. This means that if there is a shop, a planet,
a cannonball or a ship within that 10meter radius, the code will find a
new spot for the AI to spawn.

The cargo ship will still try to spawn directly in front of the player,
but if there is something in the way if will spawn randomly, just like
the Marine.