Sound effects, songs, spawn tweek
Added some sound effects to the game.
Added effects to:
Cannons being fired.
Cannonballs hitting something.
Player clicking anywhere in the shop.

There are 3 looping songs in the game:
One ambient song that constantly plays.
One that plays around shops, volume depending on distance player has to
the shop.
Ambient song when the player is in the shop.

There are now only 3 marines in the world at any given time instead of
10. I think this is a better number for now.