Tweaked boss, updated behavious when combat is started
The boss was firing to often, making him impossible to defeat. I
increased the cooldown of his weapons from 1 second to 2 seconds. This
allows the player to actually shoot back without being vaporized.

It used to be so that when a Marine/Cargo ship spotted the player, all
other ships would be deleted. This functionality was based on the idea
of having one camera for when the player drives around, that is far
abrove the player ship, and one camera for when combat is started that
is close to the player.
This is not implemented yet, and might never be, so I changed it to be
so that all other ships that are not to engage in combat will flee from
the player. This still looks a little weird, but not as weird as them
all just despawning.