Added cheats! Fixed fleeing bug
The player kan now type in cheat codes. They are pretty simple for now,
but it allows us to easily jump to and from locations during gametests.
Pressing Enter will open a window for the player to type in. Pressing
Enter again will close the window and check what was typed inside it.

There was a bug causing all enemies to flee when at 20% remaining
health. This is now fixed.