Optimizing, tweaks and bugfixes
Today I focused on bugfixes and optimization. Originally I wanted to
focus on the optimization part, but there was a lot of stuff to fix, so
I did it while I knew what was wrong.

AIMaster.cs:

Change to material 2 or 3 on the IA: This check is now done every time a
bullet or bomb hits the AI instead of in every frame.
Test if the AI should flee: moved it from Update() to every time a
bullet hits the AI, and the AI has less than 33% remaining health.
No longer checks if AI's health is equal or below 0 once every frame.
This is instead done when the AI hits a bomb or projectile.
AI that got to far away from the player would despawn, which is good,
but they would also drop scraps and reset the bools preventing new AI
from being spawned and reset the sopFightTimer, which is bad. AI will
flee and despawn when a combat is ongoing, so fucking up the bools
preventing new AI to spawn is bad. I created a new function
"killAbsentAI" that is run when an AI is to despawn due to distance, and
this function does not spawn scraps or reset bools.
Instead of first having to calculate the distance between AI and player
to determine whether or not to charge in for an attack, we now first
check if the AI already is attacking or not.

AIProjectile.cs tweaks:

Removed unused function "KillProjectile" from AIProjectile.cs
The Update() function no longer search for the GameObject called
PlayerShip once every frame. Instead we get the playership in Start(),
then use the reference to the object to get the players position.
AIavoid.cs tweaks:

Added back "hitObject" in AIavoid as this bool is crucial for the AI to
successfully navigate the world. The bool is true for as long as a
rayCast hits a shop, planet or another ship. I decided to remove this
bool some time ago, as I though it did nothing. Well both AIPatroling.cs
and AImove.cs uses this bool to navigate in the world, and the scripts
will only function if the bool is false, which it never would revert to
be.
Created bool "hitSide" wich is altered by "CastRays".cs, and is used to
check if the ai hit something to the left or right.
CastRays.cs tweaks:

Alters the bool "hitSide" in "AIavoid", and should help the AI navigate
even better in the world. Previously this script would alternate the
same bool as AIavoid would, hitObject, but I feel that maybe if this
script set it to true, AIavoid sometime could set it to false anyways.
So now both AImove and AIpatroling will function if both these bools are
false.
I forgot to say what were to happen if the raycasts hit nothing, meaning
that the AI would avoid something, and when there was nothing to avoid
the variables would not be reset. This is now fixed.
AIPatrling.cs tweaks:

Removed the calculation that would see how far away the AI is from the
target patroling point. The script would check if the AI was less than
10m away from the point, and then tell the AIMaster to deactivate
patroling and go into attack mode. I don't ever see how or when this
will be useful, seen as how all patrolpoints are far outside the map.
gravityPull.cs tweaks:

4 of the variables used in this script would be declared once every
frame, so I moved the declaration out of the Update function and
declared them when the script is made. This should save like 5 frames
every 1000 frames, by my calculations.
Removed a part of the if-test in the Update function. It would check if
the name of the RigidBody object was anything else than "moon2". moon2
does not even exist any more, so I removed this part of the test.
cameraController.cs tweaks:

3 variables would be declared one every frame. I moved the declaration
to occur once the script is made, and put the reference to the Player
and Camera objects in "Start()", so that the script won't have to search
for the object every frame and instead can use the reference to the
object.
Respawn.cs tweaks:

The array "shops" would update once every frame, when in reality we only
need to update it the instance the player is to respawn. I moved the
line of code to the "RespawnPlayer()".
General Tweaks:

AI now has a higher detect distance (increased from 60m to 100m)
AI check if they should flee at 33% remaining health (same time they
change to Mat3 material)
Boss can no longer flee.
All projectiles will be deleted after 20 seconds. We had an issue where
the cannonballs would seemingly follow the player, as they went the same
direction and had the same speed. This change is to combat that.
An AI will spawn with the same maxVelocity and force as the player has.
When a Marine or Boss discovers the player, they will get the same force
as the player + 150 until they are closer than "aiRadar - 100". This
let's the AI catch up with the player when they spot him.
Reduced Cargo ship speed when he spots the player. This lets the player
catch up with the bugger.
When an AI or the player hits a bomb they should no longer fly in
1000km/t. I worked around this by increasing the mass, drag and
angulardrag to 5 when they get hit, and after 1 second it's reset to
default. Still a flawed function.
Increased detectDistance in "CastRays.cs" from 30m to 50m. This should
help the AI spot planets and shops before they get to close to them.
General fixes:

Renamed "RayCastPlayer" to "ExtraCannonSensors" to better indicate the
function of the script.
Cleaned "setCurrentHealth.cs". For some reason the lining were messed
up, so I just put everything back where it should be.
Cleaned "shopButtons.cs" the same way as "setCurrentHealth.cs".
Cleaned "updateHealth.cs" the same way as "setCurrentHealth.cs".
Cleaned "updatePayment.cs" the same way as "setCurrentHealth.cs".