Changed AI behaviour about fleeing
A Cargo ship will only flee once hit by a projectile or bomb. Before
that it will drive towards it's target.
A Marine will flee only when a fight already is going on, and it's too
close to the player. If a cargo ship is attacked by the player and
starts fleeing, this will not count as a fight.

All AI that is fleeing will stopp fleeing once outside of the players
radar, and return to patrolling.

When the boss spawns all ai will flee, and the status is set to
"fighting".