New Minimap implemented!
I made a new minimap from the bottom, and it is now implemented and
fully functional! The issue we had with the previous one was that it
would only display a minimap icon when the object was closer than X
distance from the player, and display the WHOLE icon at the same time,
not just what's inside the minimap. This was an issue espesially with
big planets.

The new minimap is round and simpler. Simpler, as in it shows everything
that's inside the radius of the minimap. Not the entire icon, just
what's inside of it.
This is the guide I followed (there are 2 parts):
https://www.youtube.com/watch?v=ZuV9Xlt-l6g