AI works in Tutorial again, border around the world
I forgot to do a test to see if we are in the Tutorial scene, but this
is now implemented. AI works there again.

There is now a border going around the playable area, shown with the
color red. When the player drives into this field a bool is set to true
and a timer starts ticking down. This is displayed on the screen for the
player to see. When the timer reaches 0, the player dies. Driving out of
the red field resets the timer.