Ship of Dead AI will stay after death
Now when an AI dies the ship will not instantly respawn. Instead it will
continue to move until stopped (due to angular drag and so on). There
also will play an explosion sound when the ai dies.
I will now work on making the ship rotate based on the direction is was
moving during it's death.

Also updated som audio files.