Fixed AI patroling
There was a bug in the code resulting in the AI always driving forwards
unless in combat. This is because the script "CastRasys" always would
set the turn left and turn right bools to false if they did not hit
anything, overwriting the data from AIPatroling.cs. This is now fixed
and should no longer be an issue.