Updated scrap and cannonball prefabs, fixed error when a cannonball hit a dead ship still playing the explosion sound
The new scrap models are in, and when the code spawn scraps it will
randomly spawn one of them. Cannonballs did not contain a hitbox, put
one on them.
In AIProjectile.cs it would check for the length of the source clip when
it hit something, even when it should not. Meaning when it hits a ship
that is dead and still is playing it's explosion sound, that ship will
prevent the cannonball from playing it's own sound clip. Even though
this happened, the script would access it's own sound sorce and check
for the clip length, causing an error to occur. This is fixed.