Changed how AIProjectile picks random sound clip
Previously it would create the variable "int tempSound". My guess is
that this requires more data than just to pick a random number between 0
and 2 directly in the "source.clip" part of the code. So, now it will
randomly create a number in the array index instead of creating a new
variable.