Ship collisions
If any ships collide they will no longer drive through eachothers. The
calculations are made in the script "ShipHitObject", and it calculates
the difference between the ships velocity before and after it hit
something.

When it has calculated how much health is to be lost due to the impact,
it will inflict this damage to the other ship it hit. This means that if
the player rams an AI, it will damage the AI based on how much speed he
(the player) lost during the impact. Same goes for how the AI damages
the player.