First iteration of big map
You can press "M" to open and close the big map. It is literally just
the minimap camera increased in size to show the entire screen. The
script will deactivate ingame_canvas and _GUI (minimap renderTexture) to
clean the screen, and they reactivate when the minimap is closed. The
game is paused while the map is open.