Map implemened, damaged AI drives slower
There's now a big map on the screen when the player presses 'M'. This,
for now, displays only 2 map markers. We will create new markers to
display for the future.

When an AI reaches Mat2 og 3 level it will have it's force reduced by
100. This is the first step in seeing how we can make the AI appear
weaker when they have been damaged.