Modified cannonballs, crash particles, new map background, new collider when AI dies
The cannonballs are now bigger and moves faster. This is to balance them
out and make them more fun to use.

When two ships collide hard enough for at least one of them to lose
health, a spark-particle simulation will spawn at the collision
position. This looks cool.

The big map has a new background. It looks more like an old map now.

When an AI dies it has no constraints, meaning the player can push them
around as he/She sees fit. This means that the huge hitbox is an issue,
seen as how the player can make a big invisible wall betwen him and
other ai's by flipping a dead ai's ship around. Now when an ai dies it
activates the old Mesh Collider and deactivates the new.