Added new boss special weapon
The new model and animation for the boss's special weapon is
implemented. If the animation is not playing, it will go 2 seconds and
then it will play. When it's done playing the bomb will spawn behind the
player. The bomb also looks like a mine now.

If the animation is playing, a timer will tick up. This one goes up to
2.1 seconds, and then it will instantiate a far to big particle effect
on the animation position, as a way to tell the player that the bomb has
been teleported away. When the bomb spawns, which happens when the
animation is dine, a similar particle effect plays on the bombs
location.