New fleeing mechanic ,fixed world border
Instead of having the ai constantly drive to have the relative x
position between itself and the player to be 0, it will now check the
players position ever 3 seconds and drive to a target that is placed at
the complete oposite direction of the player.
In short: The ai recieves a new patroling point every 3 seconds.
This makes it a LOT easier for the player to catch up to the fleeing ai,
and is in no way near as anoying.

There was a bug causing the text on the screen not to show when the
player is out of bounds. This is fixed.