World map: Update background, new sun and planet icons
We wanted all objects on the world map to appear inside the red square
on the map, so we tried to apply a map to the texture. Sadly, this did
not work because the background simply is a object the camera renders,
and not what the camera renders on.
The "hack" around this was to use multiple background images. I did this
by stretching the minimapRendererd background to be to big for the
render plane, and then render a second background picture on the
MainCamera, that is a little bit wider than the minimap rendered one.
It now looks like all map objects are rendered inside the red square.

New icons for the sun and the planets are up and running. Player, shop,
treasure planets/ships and moons to come!